+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # splat Release Notes
 
+### 0.30.0
+
+* Add compiler options supported by spimdisasm.
+  * New compilers: `KMC` (n64), `EGCS` (iQue), `PSYQ` (PS1) and `MWCCPS2` (PS2)
+
 ### 0.29.0
 
 * Fix bss/sbss asm files not being generated.